Move DeserializingIterator to stubs package, clean-up imports
