+++ pact-scala/pact-scala-core/src/main/scala/eu/stratosphere/scala/operators/MapOperator.scala
@@ -15,21 +15,17 @@ package eu.stratosphere.scala.operators
 
 import language.experimental.macros
 import scala.reflect.macros.Context
-import eu.stratosphere.scala.codegen.MacroContextHolder
-import eu.stratosphere.scala.ScalaContract
+
 import eu.stratosphere.pact.common.contract.MapContract
-import eu.stratosphere.scala.analysis.UDT
 import eu.stratosphere.pact.common.`type`.PactRecord
 import eu.stratosphere.pact.common.stubs.{Collector, MapStub => JMapStub}
 import eu.stratosphere.pact.generic.contract.Contract
-import eu.stratosphere.scala.contracts.Annotations
-import eu.stratosphere.scala.OneInputScalaContract
-import eu.stratosphere.scala.analysis.UDF1
-import eu.stratosphere.scala.analysis.UDTSerializer
 import eu.stratosphere.nephele.configuration.Configuration
-import eu.stratosphere.scala.DataSet
-import eu.stratosphere.scala.OneInputHintable
-import eu.stratosphere.scala.codegen.Util
+
+import eu.stratosphere.scala.codegen.{MacroContextHolder, Util}
+import eu.stratosphere.scala._
+import eu.stratosphere.scala.analysis._
+import eu.stratosphere.scala.contracts.Annotations
 import eu.stratosphere.scala.stubs.{MapStub, FlatMapStub, FilterStub, MapStubBase}
 
 object MapMacros {
